# docs/swagger.yaml
basePath: /v1
definitions:
...
  forms.CreateServiceForm:
    properties:
      description:
        maxLength: 1000
        minLength: 10
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - description
    - name
    type: object
...
  models.ErrorResponse:
    properties:
      details: {}
      message:
        type: string
      traceId:
        type: string
      type:
        type: string
    type: object
...
  models.PaginatedResult-models_Service:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      meta:
        properties:
          currentPage:
            type: integer
          nextPage:
            type: integer
          totalCount:
            type: integer
          totalPages:
            type: integer
        type: object
    type: object
...
  models.Service:
    properties:
      createdAt:
        description: |-
          gorm:"<-:create" only allows create and read but not update
          this is avoid updating created_at with a zero value by mistake
        type: string
      description:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/models.ServiceMetadata'
      name:
        type: string
      organization_id:
        type: string
      updatedAt:
        type: string
    type: object
  models.ServiceMetadata:
    properties:
      versionCount:
        type: integer
    type: object
...
host: localhost:9000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server for the Konnect Platform
  termsOfService: http://swagger.io/terms/
  title: Konnect
  version: "1.0"
paths:
...
  /orgs/{orgId}/services:
    get:
      consumes:
      - application/json
      description: Gets all the services available
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Service name, supports searching the passed string in the name
          of the service
        in: query
        name: q
        type: string
      - description: Sort order for the list of services. Accepted values are asc
          and desc. Default is desc(assumes default on invalid values as well)
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - description: The field on which sorting to be applied, supports name, created_at,
          updated_at. Default is updated_at(assumes default on invalid values as well)
        enum:
        - name
        - created_at
        - updated_at
        in: query
        name: sort_by
        type: string
      - description: Page number for pagination (0-based). Default is 0
        in: query
        name: page
        type: integer
      - description: Number of items per page. Default is 10, max is 100, assumes
          100 if >100 is passed
        in: query
        name: per_page
        type: integer
      - description: 'Additional data to include (comma-separated). Supported values:
          versionCount'
        in: query
        name: include
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResult-models_Service'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All services
      tags:
      - Service
    post:
      consumes:
      - application/json
      description: Creates a service
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Service
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/forms.CreateServiceForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Service'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a service
      tags:
      - Service
  /orgs/{orgId}/services/{serviceId}:
    delete:
      consumes:
      - application/json
      description: Deletes the specified service
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Service ID
        in: path
        name: serviceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a service
      tags:
      - Service
    get:
      consumes:
      - application/json
      description: Gets the specified service
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Service ID
        in: path
        name: serviceId
        required: true
        type: string
      - description: 'Additional data to include (comma-separated). Supported values:
          versionCount'
        in: query
        name: include
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Service'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a service
      tags:
      - Service
    put:
      consumes:
      - application/json
      description: Updates the specified service
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Service ID
        in: path
        name: serviceId
        required: true
        type: string
      - description: Service
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/forms.CreateServiceForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Service'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a service
      tags:
      - Service
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"